Rework onnx import: onnxsim CF and removed torch coupling
We now use the onnxsim package for constant folding, which is helpful
when ORT is not installed (which means our CF doesn't work).

The parent_pt_model argument to the ONNX importer is also removed.
Rather than reading the parameter values from the pt models during onnx
import, the parameters are now expected as inputs to the model