Add ONNX node tests to CI, cleanup more pure ops
With the new decorator feature, we can simplify a whole bunch of ops.

Also added the test cases ONNX ships with to our CI, this is quite
useful for development. For now, we are still skipping quite a few,
especially on the more complex ops.

Pull Request: https://github.com/spcl/daceml/pull/100