Add DistributedMemlet node and scheduling function
This change adds the DistributedMemlet library node and the scheduling
function for distributed computation.

This allows you to distribute the work in the top-level map of the SDFG
by specifying block sizes. The lowering function will analyze the SDFG
and try to find MPI nodes that implement the required communication.

Pull Request: https://github.com/spcl/daceml/pull/120