Support nested SDFGs in distributed lowering
This allows us to support reductions with their intialization states.
The idea is that nested SDFG are required to be schedule such that there
is no communication within them. The user passes the schedules for each
map, and the implied communication constraints are then checked for
consistency.

Keeping communication out of the Nested SDFGs means that there is no
communication between things like reduction buffer initialization, and
also means that all global communication is kept top-level, where it is
easier to optimize

Pull Request: https://github.com/spcl/daceml/pull/123